Correct formatting of keywords.txt
Each field of keywords.txt is separated by a single true tab. When you use multiple tabs it causes the field to be interpreted as empty. On Arduino IDE 1.6.5 and newer an empty KEYWORD_TOKENTYPE causes the default editor.function.style highlighting  to be used (as with KEYWORD2, KEYWORD3). On Arduino IDE 1.6.4 and older it causes the keyword to not be recognized for any special highlighting.

Some of the keywords identifiers were incorrectly spelled "KEYW0RDS" (with a zero) rather than "KEYWORDS" (with the letter o).

The keyword identifier for the readRegister16 keyword definition was missing, causing it to not be highlighted by the Arduino IDE.

Reference:
https://github.com/arduino/Arduino/wiki/Arduino-IDE-1.5:-Library-specification#keywords